Address comments from review
